+++ tests/test_cli_export_snippet.py
@@ -35,7 +35,7 @@ from tests.testlib.snippet import Snippet
 class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
     """Test workflows for exporting snippets."""
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_001(self, snippy):
         """Export all snippets.
 
@@ -53,9 +53,9 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_002(self, snippy):
         """Export all snippets.
 
@@ -73,7 +73,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export', '--snippets'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
     @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
     def test_cli_export_snippet_003(self, snippy):
@@ -110,9 +110,9 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
             ]
         }
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
-            cause = snippy.run(['snippy', 'export', '-f', './defined-snippets.yaml', '--snippet'])
+            cause = snippy.run(['snippy', 'export', '-f', './defined-snippets.yml', '--snippet'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './defined-snippets.yaml', content)
+            Content.assert_yaml(yaml, mock_file, './defined-snippets.yml', content)
 
     @pytest.mark.usefixtures('default-snippets')
     def test_cli_export_snippet_005(self, snippy):
@@ -130,7 +130,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
             file_handle = mock_file.return_value.__enter__.return_value
             file_handle.write.assert_not_called()
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_006(self, snippy):
         """Export defined snippets.
 
@@ -148,7 +148,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export', '-d', '53908d68425c61dc'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
     @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
     def test_cli_export_snippet_007(self, snippy):
@@ -221,7 +221,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
             assert cause == 'NOK: cannot find content with message digest: 123456789abcdef0'
             Content.assert_text(mock_file, None, None)
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_011(self, snippy):
         """Export defined snippets.
 
@@ -239,7 +239,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export', '--sall', 'force'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
     @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
     def test_cli_export_snippet_012(self, snippy):
@@ -350,7 +350,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
             assert cause == 'NOK: cannot find content with given search criteria'
             mock_file.assert_not_called()
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_018(self, snippy):
         """Export defined snippet with content data.
 
@@ -368,7 +368,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export', '--content', 'docker rm --volumes $(docker ps --all --quiet)'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
     @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
     def test_cli_export_snippet_019(self, snippy):
@@ -492,7 +492,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
             assert cause == 'NOK: no content found to be exported'
             mock_file.assert_not_called()
 
-    @pytest.mark.usefixtures('yaml', 'default-snippets', 'export-time')
+    @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_026(self, snippy):
         """Export snippets with search keyword.
 
@@ -511,7 +511,7 @@ class TestCliExportSnippet(object):  # pylint: disable=too-many-public-methods
         with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
             cause = snippy.run(['snippy', 'export', '--sall', 'docker'])
             assert cause == Cause.ALL_OK
-            Content.assert_yaml(yaml, mock_file, './snippets.yaml', content)
+            Content.assert_mkdn(mock_file, './snippets.mkdn', content)
 
     @pytest.mark.usefixtures('default-snippets', 'export-time')
     def test_cli_export_snippet_027(self, snippy):